requirements.txt
charset-normalizer==3.3.2
Cython==3.0.9
idna==3.6
numpy==1.21.6
opencv-python==4.9.0.80
pygame==2.5.2
PyOpenGL 
PyOpenGL-accelerate 
pyzed
requests==2.31.0
urllib3==2.0.7
wincertstore==0.2
